requirements.txt
django
requests
psycopg2
djangorestframework
djangorestframework-simplejwt
